broke up server.cpp into multiple files
