+++ daemon/src/main/native/server/server.h
