+++ test-resources/.env.mock
@@ -1,13 +1,13 @@
 WF_DEVICE_NAME="test"
-WF_TEAM="1234"
-WF_SLAM_SERVER="true"
+WF_TEAM=1234
+WF_SLAM_SERVER=true
 WF_LOCAL="../../../test-resources/mock_local"
 WF_RESOURCE="../../../resources"
-WF_API_PORT="8080"
+WF_API_PORT=8080
 WF_MODELS_RSUBDIR="models"
 WF_FIELDS_RSUBDIR="fields"
 WF_HARDWARE_LSUBDIR="hardware"
 WF_PIPELINE_LSUBDIR="pipelines"
-WF_EXEC=""
+WF_EXEC=
 WF_SCRIPTS="../../../scripts"
